requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==1.26.4
opencv-python==4.9.0.80
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
ultralytics
paho-mqtt
flask
flask-socketio
eventlet
